feat: 스터디 참가 취소하기-카프카
