+++ src/main/java/com/suite/suite_suite_room_service/suiteRoom/service/SuiteRoomServiceImpl.java
@@ -83,7 +83,7 @@ public class SuiteRoomServiceImpl implements SuiteRoomService{
             throw new CustomException(StatusCode.FAILED_PAY);
 
         suiteRoomRepository.save(suiteRoom);
-        suiteRoomProducer.sendPaymentMessage(suiteRoom, authorizerDto, true);
+        suiteRoomProducer.sendPaymentMessage(suiteRoom, authorizerDto, true, true);
 
 
     }
